+++ libpolys/coeffs/ffields.h
@@ -13,7 +13,7 @@ BOOLEAN nfInitChar(coeffs r, void*);
 
 BOOLEAN nfGreaterZero (number k, const coeffs r);
 number  nfMult        (number a, number b, const coeffs r);
-number  nfInit        (int i, const coeffs r);
+number  nfInit        (long i, const coeffs r);
 number  nfPar         (int i, const coeffs r);
 int     nfParDeg      (number n, const coeffs r);
 int     nfInt         (number &n, const coeffs r);
