+++ libpolys/polys/ext_fields/transext.h
@@ -104,7 +104,7 @@ BOOLEAN  ntEqual(number a, number b, const coeffs cf);
 BOOLEAN  ntIsOne(number a, const coeffs cf);
 BOOLEAN  ntIsMOne(number a, const coeffs cf);
 BOOLEAN  ntIsZero(number a, const coeffs cf);
-number   ntInit(int i, const coeffs cf);
+number   ntInit(long i, const coeffs cf);
 int      ntInt(number &a, const coeffs cf);
 number   ntNeg(number a, const coeffs cf);
 number   ntInvers(number a, const coeffs cf);
